renaming the tabs context
