+++ renderer/App/App.js
@@ -2,7 +2,7 @@ const { html, css } = require('../tools/ui.js');
 const { withConfig } = require('../tools/config.js');
 const { Queue, withQueue } = require('../Queue/Queue.js');
 const { withTheme } = require('../tools/theme.js');
-const { withTabs, useTabs } = require('../TabBar/TabBar.js');
+const { withTabs, useTabs } = require('../Tabs/Tabs.js');
 
 const Frame = require('../Frame/Frame.js');
